Squashed commit of the following:
commit 456f8bf2b6d0f37579c46df8d38063da4529892f
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Mon Jan 13 01:43:17 2025 -0800

    cleaned up frontend styling; added logo-spinner for loading

commit b0ad07c7aeadd369b9f2f40a5b81c39ab3689a79
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sun Jan 12 23:47:42 2025 -0800

    recipe detail route and frontend skeleton

commit c69219ca73654c0ec7f90de0604ec6ee42592d87
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sun Jan 12 23:20:14 2025 -0800

    all recipes front end

commit 9a69415bf6b0ceb7369f55d6b634a3722d6a1450
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sun Jan 12 19:06:32 2025 -0800

    updated ingredients and instructions fields, completed create recipe, updated get all recipes and get all recipes by owner_id

commit ecd26efd5fb8cde5c874f05a955bdd217a95af46
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sat Jan 11 17:19:40 2025 -0800

    create recipe wip

commit 43e4b725b00c7bfc84d9d2955c40c0b7bc3add1e
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sat Jan 11 14:48:52 2025 -0800

    get restaurants by owner id

commit 66ae731984f93567cc4f22528a4db9fc2c80f744
Merge: c17b427 a639882
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sat Jan 11 13:59:32 2025 -0800

    updating delete user profile to follow RESTful principles

commit a6398825533497cc7961429dbdf01b4cba5ff4dc
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sat Jan 11 13:00:27 2025 -0800

    updated login and sign up error handling routes and api docs; updated seed data with ratatouille refs

commit c17b4274b2b01b6c772c7d8b37baaf11381ed36f
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Sat Jan 11 13:00:27 2025 -0800

    updated login and sign up error handling routes and api docs; updated seed data with ratatouille refs

commit 6117505fc16f3c8bdc5d367348f8898278d1d56b
Merge: 85f6212 70eb838
Author: lainey1 <lainey.d.paddock@gmail.com>
Date:   Fri Jan 10 17:51:10 2025 -0800

    Merge remote-tracking branch 'origin/feature/user-profiles' into dev

+++ app/api/user_routes.py
@@ -23,7 +23,7 @@ def user(id):
     Query for a user by id and returns that user in a dictionary
     """
     user = User.query.get(id)
-    return user.to_dict()
+    return {'user': user.to_dict()}
 
 @user_routes.route('/<int:user_id>', methods=['PUT'])
 @login_required
@@ -41,7 +41,10 @@ def update_profile(user_id):
 
     # if user is not the profile owner, return 403 Forbidden
     if user.id != current_user.id:
-        return jsonify({'message': 'Only the profile owner can edit'}), 403
+        return jsonify({
+        "message": "Only the profile owner can make changes to these details."
+        }
+        ), 403
 
 
     # if profile owner, create instance of UserProfileForm class
@@ -77,35 +80,42 @@ def update_profile(user_id):
         'errors': form.errors
     }), 400
 
-
-@user_routes.route('/delete_profile', methods=["GET", "POST"])
+@user_routes.route('/<int:user_id>', methods=["POST"])
 @login_required
-def delete_profile():
-    """
-    Delete the user's profile by user ID.
-    """
+def delete_profile(user_id):
+    """Delete the user's profile by user ID."""
+    if user_id != current_user.id:
+        return jsonify({"message": "Only the profile owner can delete their account."}), 403
+
+    # if profile owner, create instance of DeleteProfileForm class
     form = DeleteUserForm()
 
     # manually sets the CSRF token in the form from the request's cookies.
     form['csrf_token'].data = request.cookies['csrf_token']
 
     if form.validate_on_submit():
-        print("Delete action triggered")  # Debugging output
-        if form.delete.data:
-            # Get the current user
-            user_to_delete = User.query.get(current_user.id)
-
-
-            # Delete the user from the database
+        if form.delete.data:  # If the "Delete Profile" button was pressed
+            user_to_delete = User.query.get(user_id)
             if user_to_delete:
                 db.session.delete(user_to_delete)
                 db.session.commit()
-
+                return jsonify({'message': 'User profile deleted successfully'}), 200
                 # return redirect(url_for('home'))  ##when available Redirect to home or login page after deletion
+        elif form.cancel.data:  # If the "Cancel" button was pressed
+            return redirect(url_for('users.user', id=current_user.id))
 
-        # elif form.cancel.data:
-        #     ## when available Redirect to the user's profile page if they cancel the deletion
-        #     return redirect(url_for('users.user', id=current_user.id))
+    return render_template("delete_profile.html", form=form)
+
+
+
+@user_routes.route('/delete_profile', methods=["GET", "POST"])
+@login_required
+def get_delete_profile_form():
+    """
+    Delete the user's profile by user ID.
+    """
+
+    form = DeleteUserForm()
 
     # print("Form errors:", form.errors)  # Debugging output
     return render_template("delete_profile.html", form=form)
